Rename music album to music
