--- lib/data/music_album.json
@@ -1,1 +1,0 @@
-["Thriller","2003-12-15","true","Classic","Elvis","Presley","Google Play","Warner","blue"]
