+++ lib/music.rb
