+++ spec/spec_helper.rb
@@ -4,7 +4,7 @@ require_relative '../lib/author'
 require_relative '../lib/source'
 require_relative '../lib/label'
 require_relative '../lib/book'
-require_relative '../lib/music_album'
+require_relative '../lib/music'
 require_relative '../lib/movie'
 require_relative '../lib/game'
 require_relative '../lib/app'
